Fix bugs with Shift/Control modifier keys on SIII
Shift and Control on the USB keyboard now have the proper behavior on SIII.
Removed some debug code that had been left in.
Fixed UART definitions, wrong mapping was accidentally checked in last time.
Increased version to 1.4.

More details:
On a Series III system, holding down Shift or Control should change the icons displayed at the bottom of the CMI screen. These icons map onto the F1-F15 function keys, and so the function of a each key changes if Shift or Control is also pressed.
This wasn't working before.
Special thanks for Graham Johnson for reporting this bug!